[ansible] xscreensaver is not enabled in VM's
The xscreensaver role will use facts to decide if it should be initially
configured with a timeout or not.

On bare metal installations it will suspend and lock the screen after 5
minutes for added security. In VM's it will be initially disabled,
since it is likely that the host has its own screensaver set up.