[vagrant] Wait for reflector to write a new mirrorlist
Sometimes reflector fails to update the mirrors, and the currently
synced DB's can be out of sync with the first mirror in the list.
As a workaround, the first sysupgrade forces a repo sync (-Syy). When
reflector does its job, it is harmless as new repos are likely to be
downloaded anyway.