[ansible] Rename `partitioning` role to `disksetup`
Current Ansible versions have a bug: if a role includes a collection role with the same name
(i.e. `FooRole` includes `mynamespace.mycollection.FooRole`) Ansible will include the global role,
not the one in the collection, causing endless recursion.

This appears to have been fixed in the currently unreleased 2.10 development branch of Ansible.