[ansible] Add optional workaround for unwanted account locking via xsecurelock
If pam_unix returns a real authentication error, let pam_faillock know
so it can decrement its counters. But for other errors, just stop at
pam_unix.

This is a workaround to a problem with xsecurelock and PAM: if the user
triggers the password prompt, then let it timeout, this will count as a
failure and decrement the lock counter, even if no password was actually
typed.