[ansible] Update theme settings
Wallpapers require a new property under recent Xfce versions.

Unfortunately, it seems this property now depends on the exact GPU
output name, such as monitorHDMI-1 instead of monitor0. To recover it,
we must actually run X and query it. The primary screen is then used as
the monitor name.

Currently, multi-screen support is not there.